requirements.txt
tensorflow==2.14.1
keras==2.14.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
matplotlib~=3.9.0
scikit-learn~=1.5.0